dash-indeed-scrap/requirements.txt
dash
gunicorn==19.9.0
stats==0.1.2a
scipy==1.3.0
dash.dependencies
dash_html_components
dash_core_components
plotly.graph_objects
numpy
Pandas
dash_table
json